third_party/WebKit/LayoutTests/editing/execCommand/justify-right-crash.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
// This is a test for crbug.com/371139
test(()=> assert_selection(
  [
  '<div contenteditable>',
    '^<textarea></textarea>',
    '<span></span>',
    '<div style="display:inline-table">foo',
      '<body>',
        '<textarea></textarea>',
        '<span></span>',
        '<div id="target" style="display:inline-table">',
          'foo|',
        '</div>',
      '</body>',
    '</div>',
  '</div>'
  ],
  "JustifyRight",
  [
  '<div contenteditable>',
    '<div style="text-align: right;">',
      '^<textarea></textarea>',
      '<div style="display: inline-table;">',
        'foo',
        '<textarea></textarea>',
        '<div id="target" style="display:inline-table">',
          'foo|',
        '</div>',
      '</div>',
    '</div>',
  '</div>']),
  "This JustifyRight the entire selected content.");
</script>
